Added MenuStructure usage in MenuRepository

+++ src/Webaccess/WCMSLaravel/Repositories/EloquentMenuRepository.php
@@ -2,8 +2,8 @@
 
 namespace Webaccess\WCMSLaravel\Repositories;
 
-use CMS\Entities\Menu;
-use CMS\Entities\MenuItem;
+use CMS\Structures\MenuStructure;
+use CMS\Structures\MenuItemStructure;
 use CMS\Repositories\MenuRepositoryInterface;
 use Webaccess\WCMSLaravel\Models\Menu as MenuModel;
 use Webaccess\WCMSLaravel\Models\MenuItem as MenuItemModel;
@@ -21,29 +21,7 @@ class EloquentMenuRepository implements MenuRepositoryInterface {
     public function findByIdentifier($identifier)
     {
         $menuDB = MenuModel::where('identifier', '=', $identifier)->first();
-
-        if ($menuDB) {
-            $menu = new Menu();
-            $menu->setIdentifier($menuDB->identifier);
-            $menu->setName($menuDB->name);
-
-            if ($menuDB->items) {
-                foreach ($menuDB->items->sortBy('order') as $itemDB) {
-                    $item = new MenuItem();
-                    $item->setLabel($itemDB->label);
-                    $item->setOrder($itemDB->order);
-                    if ($itemDB->page_id) {
-                        $page = PageModel::find($itemDB->page_id);
-                        $item->setPage($this->pageRepository->findByIdentifier($page->identifier));
-                    }
-                    $menu->addItem($item);
-                }
-            }
-
-            return $menu;
-        }
-        
-        return false;
+        return ($menuDB) ? $this->convertMenuModelToMenuStructure($menuDB) : false;
     }
 
     public function findAll()
@@ -52,52 +30,74 @@ class EloquentMenuRepository implements MenuRepositoryInterface {
 
         $menus = [];
         foreach ($menusDB as $i => $menuDB) {
-            $menu = new Menu();
-            $menu->setIdentifier($menuDB->identifier);
-            $menu->setName($menuDB->name);
-
-            foreach ($menuDB->items as $itemDB) {
-                $item = new MenuItem();
-                $item->setLabel($itemDB->label);
-                $item->setOrder($itemDB->order);
-                if ($itemDB->page_id) {
-                    $page = PageModel::find($itemDB->page_id);
-                    $item->setPage($this->pageRepository->findByIdentifier($page->identifier));
-                }
-                $menu->addItem($item);
-            }
-        
-            $menus[]= $menu;
+            $menus[]= $this->convertMenuModelToMenuStructure($menuDB);
         }
 
         return $menus;
     }
 
-    public function createMenu(Menu $menu)
+    public function createMenu(MenuStructure $menuStructure)
     {
-        $menuDB = new MenuModel();
-        $menuDB->identifier = $menu->getIdentifier();
-        $menuDB->name = $menu->getName();
-
+        $menuDB = $this->convertMenuStructureToMenuModel($menuStructure, new MenuModel());
         return $menuDB->save();
     }
 
-    public function updateMenu(Menu $menu)
+    public function updateMenu(MenuStructure $menuStructure)
     {
-        $menuDB = MenuModel::where('identifier', '=', $menu->getIdentifier())->first();
-        $menuDB->name = $menu->getName();
+        $menuDB = MenuModel::where('identifier', '=', $menuStructure->identifier)->first();
+        $menuDB = $this->convertMenuStructureToMenuModel($menuStructure, $menuDB);
+        
+        return $menuDB->save();
+    }
+
+    public function deleteMenu(MenuStructure $menuStructure)
+    {
+        $menuDB = MenuModel::where('identifier', '=', $menuStructure->identifier)->first();
+        
+        return $menuDB->delete();
+    }
+
+    public function convertMenuModelToMenuStructure(MenuModel $menuModel)
+    {
+        $menuStructure = new MenuStructure();
+        $menuStructure->identifier = $menuModel->identifier;
+        $menuStructure->name = $menuModel->name;
+        $menuStructure->items = [];
+
+        if ($menuModel->items) {
+            foreach ($menuModel->items->sortBy('order') as $itemDB) {
+                $item = new MenuItemStructure();
+                $item->label = $itemDB->label;
+                $item->order = $itemDB->order;
+                if ($itemDB->page_id) {
+                    $page = PageModel::find($itemDB->page_id);
+                    $item->page = $page->identifier;
+                }
+                $menuStructure->items[]= $item;
+            }
+        }
+
+        return $menuStructure;
+    }
+
+    public function convertMenuStructureToMenuModel(MenuStructure $menuStructure, $menuDB)
+    {
+        $menuDB->name = $menuStructure->name;
 
         //Delete existing items
-        MenuItemModel::where('menu_id', '=', $menuDB->id)->delete();
+        if ($menuDB->id)
+            MenuItemModel::where('menu_id', '=', $menuDB->id)->delete();
+        else
+            $menuDB->identifier = $menuStructure->identifier;
         
         //Add new items
-        if (is_array($menu->getItems())) {
-            foreach ($menu->getItems() as $item) {
+        if (is_array($menuStructure->items)) {
+            foreach ($menuStructure->items as $item) {
                 $itemDB = new MenuItemModel();
-                $itemDB->label = $item->getLabel();
-                $itemDB->order = $item->getOrder();
-                if ($item->getPage()) {
-                    $pageDB = PageModel::where('identifier', '=', $item->getPage()->getIdentifier())->first();
+                $itemDB->label = $item->label;
+                $itemDB->order = $item->order;
+                if ($item->page) {
+                    $pageDB = PageModel::where('identifier', '=', $item->page->identifier)->first();
                     $itemDB->page_id = $pageDB->id;
                 }
                 $itemDB->menu_id = $menuDB->id;
@@ -105,15 +105,8 @@ class EloquentMenuRepository implements MenuRepositoryInterface {
                 $menuDB->items()->save($itemDB);
             }
         }
-        
-        return $menuDB->save();
-    }
 
-    public function deleteMenu(Menu $menu)
-    {
-        $menuDB = MenuModel::where('identifier', '=', $menu->getIdentifier())->first();
-        
-        return $menuDB->delete();
+        return $menuDB;
     }
     
 }
